@@ -1,6 +1,6 @@
 {
   "name": "meetingsdk-sample-react",
-  "version": "3.1.6",
+  "version": "3.5.0",
   "author": "Zoom Video Communications, Inc.",
   "contributors": [
     {
@@ -10,12 +10,11 @@
   "homepage": "",
   "private": true,
   "dependencies": {
-    "@zoom/meetingsdk": "^3.1.6",
-    "react": "16.14.0",
-    "react-dev-utils": "^12.0.1",
-    "react-dom": "16.14.0",
+    "@zoom/meetingsdk": "^3.5.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-scripts": "^5.0.1",
-    "web-vitals": "^3.5.0"
+    "web-vitals": "^3.5.2"
   },
   "scripts": {
     "start": "react-scripts start",
